requirements.txt
pandas>=2.0
numpy>=1.23
redis>=4.5 # includes redis.asyncio
pydantic>=2.0
python-dotenv>=1.0
aiohttp>=3.8
websockets>=11.0
asyncpg>=0.29.0
fastapi>=0.104.0
uvicorn>=0.24.0 